requirements.txt
alembic==1.11.1
anyio==3.7.0
astroid==2.15.5
asyncpg==0.27.0
black==23.3.0
certifi==2023.5.7
click==8.1.3
coverage==7.2.7
dill==0.3.6
dnspython==2.3.0
email-validator==2.0.0.post2
Faker==18.11.2
fastapi==0.97.0
greenlet==2.0.2
h11==0.14.0
httpcore==0.17.2
httptools==0.5.0
httpx==0.24.1
idna==3.4
iniconfig==2.0.0
isort==5.12.0
itsdangerous==2.1.2
Jinja2==3.1.2
lazy-object-proxy==1.9.0
Mako==1.2.4
MarkupSafe==2.1.3
mccabe==0.7.0
mypy-extensions==1.0.0
orjson==3.9.1
packaging==23.1
pathspec==0.11.1
platformdirs==3.6.0
pluggy==1.2.0
psycopg2-binary==2.9.6
pydantic==1.10.9
pytest==7.4.0
python-dateutil==2.8.2
python-dotenv==1.0.0
python-multipart==0.0.6
PyYAML==6.0
ruff==0.0.277
six==1.16.0
sniffio==1.3.0
SQLAlchemy==2.0.16
starlette==0.27.0
tomlkit==0.11.8
typing_extensions==4.6.3
ujson==5.8.0
uvicorn==0.22.0
uvloop==0.17.0
watchfiles==0.19.0
websockets==11.0.3
wrapt==1.15.0
